+++ feature/timeline/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/feature/timeline/TimelineScreen.kt
@@ -23,8 +23,6 @@ import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.pullrefresh.PullRefreshIndicator
 import androidx.compose.material.pullrefresh.pullRefresh
 import androidx.compose.material.pullrefresh.rememberPullRefreshState
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.HorizontalDivider
@@ -66,6 +64,7 @@ import com.livefast.eattrash.raccoonforfriendica.core.commonui.components.Custom
 import com.livefast.eattrash.raccoonforfriendica.core.commonui.components.ListLoadingIndicator
 import com.livefast.eattrash.raccoonforfriendica.core.commonui.components.di.getFabNestedScrollConnection
 import com.livefast.eattrash.raccoonforfriendica.core.commonui.content.ConfirmMuteUserBottomSheet
+import com.livefast.eattrash.raccoonforfriendica.core.commonui.content.CustomConfirmDialog
 import com.livefast.eattrash.raccoonforfriendica.core.commonui.content.OptionId
 import com.livefast.eattrash.raccoonforfriendica.core.commonui.content.PollVoteErrorDialog
 import com.livefast.eattrash.raccoonforfriendica.core.commonui.content.TimelineItem
@@ -435,39 +434,13 @@ class TimelineScreen : Screen {
         }
 
         if (confirmDeleteEntryId != null) {
-            AlertDialog(
-                onDismissRequest = {
+            CustomConfirmDialog(
+                title = LocalStrings.current.actionDelete,
+                onClose = { confirm ->
+                    val entryId = confirmDeleteEntryId ?: ""
                     confirmDeleteEntryId = null
-                },
-                title = {
-                    Text(
-                        text = LocalStrings.current.actionDelete,
-                        style = MaterialTheme.typography.titleMedium,
-                    )
-                },
-                text = {
-                    Text(text = LocalStrings.current.messageAreYouSure)
-                },
-                dismissButton = {
-                    Button(
-                        onClick = {
-                            confirmDeleteEntryId = null
-                        },
-                    ) {
-                        Text(text = LocalStrings.current.buttonCancel)
-                    }
-                },
-                confirmButton = {
-                    Button(
-                        onClick = {
-                            val entryId = confirmDeleteEntryId ?: ""
-                            confirmDeleteEntryId = null
-                            if (entryId.isNotEmpty()) {
-                                model.reduce(TimelineMviModel.Intent.DeleteEntry(entryId))
-                            }
-                        },
-                    ) {
-                        Text(text = LocalStrings.current.buttonConfirm)
+                    if (confirm && entryId.isNotEmpty()) {
+                        model.reduce(TimelineMviModel.Intent.DeleteEntry(entryId))
                     }
                 },
             )
@@ -500,52 +473,27 @@ class TimelineScreen : Screen {
 
         if (confirmBlockEntry != null) {
             val creator = confirmBlockEntry?.reblog?.creator ?: confirmBlockEntry?.creator
-            AlertDialog(
-                onDismissRequest = {
+            CustomConfirmDialog(
+                title =
+                    buildString {
+                        append(LocalStrings.current.actionBlock)
+                        val handle = creator?.handle ?: ""
+                        if (handle.isNotEmpty()) {
+                            append(" @$handle")
+                        }
+                    },
+                onClose = { confirm ->
+                    val entryId = confirmBlockEntry?.id
                     confirmBlockEntry = null
-                },
-                title = {
-                    Text(
-                        text =
-                            buildString {
-                                append(LocalStrings.current.actionBlock)
-                                val handle = creator?.handle ?: ""
-                                if (handle.isNotEmpty()) {
-                                    append(" @$handle")
-                                }
-                            },
-                        style = MaterialTheme.typography.titleMedium,
-                    )
-                },
-                text = {
-                    Text(text = LocalStrings.current.messageAreYouSure)
-                },
-                dismissButton = {
-                    Button(
-                        onClick = {
-                            confirmBlockEntry = null
-                        },
-                    ) {
-                        Text(text = LocalStrings.current.buttonCancel)
-                    }
-                },
-                confirmButton = {
-                    Button(
-                        onClick = {
-                            val entryId = confirmBlockEntry?.id
-                            val creatorId = creator?.id
-                            confirmBlockEntry = null
-                            if (entryId != null && creatorId != null) {
-                                model.reduce(
-                                    TimelineMviModel.Intent.BlockUser(
-                                        userId = creatorId,
-                                        entryId = entryId,
-                                    ),
-                                )
-                            }
-                        },
-                    ) {
-                        Text(text = LocalStrings.current.buttonConfirm)
+                    val creatorId = creator?.id
+                    confirmBlockEntry = null
+                    if (confirm && entryId != null && creatorId != null) {
+                        model.reduce(
+                            TimelineMviModel.Intent.BlockUser(
+                                userId = creatorId,
+                                entryId = entryId,
+                            ),
+                        )
                     }
                 },
             )
